+++ dnnv/logging/__init__.py
@@ -4,6 +4,7 @@ import argparse
 import logging
 import os
 import sys
+import tensorflow.compat.v1 as tf
 
 from contextlib import contextmanager
 from functools import partial
@@ -53,12 +54,15 @@ def initialize(name: str, args: argparse.Namespace):
         logger.setLevel(logging.DEBUG)
         suppress = partial(suppress, filter_level=logging.DEBUG)
     elif args.verbose:
+        tf.logging.set_verbosity(tf.logging.ERROR)
         logger.setLevel(logging.INFO)
         suppress = partial(suppress, filter_level=logging.INFO)
     elif args.quiet:
+        tf.logging.set_verbosity(tf.logging.ERROR)
         logger.setLevel(logging.ERROR)
         suppress = partial(suppress, filter_level=logging.ERROR)
     else:
+        tf.logging.set_verbosity(tf.logging.ERROR)
         logger.setLevel(logging.INFO)
         suppress = partial(suppress, filter_level=logging.INFO)
