+++ dnnv/verifiers/eran/errors.py
@@ -1,5 +1,9 @@
 from dnnv.verifiers.common import VerifierError, VerifierTranslatorError
 
 
+class ERANError(VerifierError):
+    pass
+
+
 class ERANTranslatorError(VerifierTranslatorError):
     pass
